requirements.txt
SQLAlchemy>=2.0.40
PyQt5>=5.15.11
PyQt5-sip>=12.17.0
Pillow>=8.0.0
psycopg2-binary>=2.9.3
pyinstaller>=6.14.1